View Code Separation and Refactoring
Separated the view code from the view controller by moving UI-related code into a dedicated CheckoutRootView. This approach avoids using storyboards and organizes the UI elements into a SwiftUI view for cleaner code management.
